backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
aiofiles
PyMuPDF
qdrant-client
couchdb
openai
numpy
python-dotenv